Rework client/server relationship
Instead of using a socketpair, use unix domain sockets. This allows us
to easily detect when the server is ready to perform the "outcome" (kexec,
reboot, poweroff, etc) since the client will get an epoll notification
when the server closes the connection and the client's subsequent read()
will return zero bytes.