Big refactor
Stop taking responsibility for doing end-user integrations of this
kernel/initrd combo into the final firmware images (e.g. coreboot). By
doing this, we can potentially be more useful by not being opinionated
on what type of firmware we are being used in. This brings the
opportunity for use in stuff like U-Boot SPL, ATF, RPI, and really
anything that can run a linux kernel.