Add EFI stub for booting tinyboot kernel/initrd
This is similar to systemd-stub, an EFI payload that is capable of
introspecting itself in order to obtain a linux/initrd pair. This can be
used to replace the BDS DXE (Boot Device Selection) if one owns the UEFI
firmware on their platform, or alternatively can just be used as the
default boot path (i.e. `EFI/boot/bootx64.efi`).